+++ reg.h
@@ -5,17 +5,17 @@
 
 typedef struct {
     uint16_t pc;
-    uint8_t r0;
-    uint8_t r1;
-    uint8_t r2;
-    uint8_t r3;
-    uint8_t r4;
-    uint8_t r5;
-    uint8_t r6;
-    uint8_t r7;
+    uint16_t r0;
+    uint16_t r1;
+    uint16_t r2;
+    uint16_t r3;
+    uint16_t r4;
+    uint16_t r5;
+    uint16_t r6;
+    uint16_t r7;
 } RegFile;
 
-void move_reg(uint8_t value, uint8_t reg, RegFile *rf);
-uint8_t get_reg(uint8_t reg, const RegFile *rf);
+void move_reg(uint16_t value, uint8_t reg, RegFile *rf);
+uint16_t get_reg(uint8_t reg, const RegFile *rf);
 
 #endif
